Prefix commands with "pubtools-pulp-"
Without a prefix, command names such as "garbage-collect" are too
generic and could easily conflict with something else in PATH.
Make sure the command names are prefixed with the project name so
we don't have to worry about such conflicts.

It's unfortunate how long the resulting command names end up being,
but as these commands are intended for use by our developers and
not our end-users, it should not be a major issue.

Note: I left the prefix off in the title of the doc page, as I
don't see much point in repeating that there for every single
command name, as long as the prefix is there in the reference
and examples themselves.