tests: exclude pubtools/__init__.py from coverage
Coverage dropped from 100% when this file was added, as the
file (at least sometimes) is measured as not covered.

This occurs due to the way the pubtools module is split across
projects. __init__.py for pubtools is provided by this project,
and pubtools-pulplib, and (in future) various other projects.
Only one of them will be loaded, and it's undefined which one.

That all works correctly, but it means that the file in this
repo isn't necessarily loaded while running this repo's tests.
I think that's OK, so let's just mark it excluded from coverage.