Introduce publish command
This patch adds a new publish command that publishes
pulp repos to the defined endpints in the repo.

The command provides filters to fetch repos in pulp
or the repos provided as input given they are present
in pulp.

The repos can be published with publish options 'force'
to publish all the content and 'clear' to remove
content

The patch includes the unit tests and refernce doc