Constraints from dependencies are included for Python 2.7
There are some constraints applied on the dependencies of dependecnies
e.g. jsonschema in pubtools-pulplib via constraints.txt. However, those
constraints are not followed here and the installation for Python 2.7
fails. Hence, addind a constraints.txt file to include the constraints
form the dependencies and using that file in tox.ini for Python 2.7
installation.