Fix pylint failures on newest pylint/astroid
An upgrade of astroid package appeared to introduce a few false
positives along the lines of:

  pubtools/_pulp/tasks/publish.py:115:23:
    E1101: Instance of 'Client' has no 'flush_repo' member (no-member)

It appears to be triggered by the way we've used multiple inheritance
in a few classes. Perhaps there's a proper way to make it work, but
in the interest of unblocking CI, let's suppress the false positives.