Introduce --pulp-fake option for development/testing
pubtools-pulplib comes with a full-featured fake client which implements
all the same API as a real client, but with a simple in-memory
implementation instead of calling out to a real server.

Enable the usage of this client for all the tasks in pubtools-pulp.
This is never expected to be used in production, but can be useful for
development as it removes the hard dependency on a real writable
Pulp server in order to realistically exercise the code within this
project.