Bump pulplib to 2.17.0, fix tests
Use pubtools-pulplib >= 2.17.0 which contains many new features and
important fixes. This version bump is being done manually because one
test needs an update.

The test update is needed because the fake in pulplib has become
more accurate in this respect: now, when units are being removed from a
repo, the unit data in the returned Task objects contains only those
fields which are present in the unit_key, same as a real server. This
means, in the case of removing RPMs from a repo, the data returned by
Pulp tells us sha256 checksum (which is part of the key) but not other
fields such as md5 and signing_key.

The test behavior now shows more accurately the way this code will
behave when run against a real Pulp server.

The relevant code under test was also slightly adjusted to avoid dead
code.