Add persistence to --pulp-fake
This change makes the --pulp-fake argument much more useful by:

- having it save and load state as each task is run
- having it seed a bit of common state (e.g. existence of
  all-rpm-content repo) if there is no prior state

The intent is to make running the tasks against a fake usable in
situations where it previously wasn't, and also make it possible to
view the changes made by a task. It's similar in concept to
developing a service against a local sqlite DB while a remote
DB is used in production.

As an example, if I run the maintenance-on command like this:

    $ pubtools-pulp-maintenance-on --pulp-fake --repo-ids myrepo
    [INFO    ] Get maintenance report: started
    [WARNING ] Using a fake Pulp client, no changes will be made to Pulp!
    [INFO    ] Get maintenance report: finished
    [INFO    ] Adjust maintenance report: started
    [WARNING ] Didn't find following repositories:
    [WARNING ]  - myrepo
    [INFO    ] Adjust maintenance report: finished
    [INFO    ] Set maintenance report: started
    [INFO    ] Uploading repos.json to redhat-maintenance [82e2e662-f728-b4fa-4248-5e3a0a5d2f34]
    [INFO    ] Set maintenance report: finished
    [INFO    ] Fake pulp state persisted to /home/rmcgover/.config/pubtools-pulp/fake.yaml

I can check the state file to see that it uploaded a repos.json into
redhat-maintenance as claimed:

    $ tail -n8 /home/rmcgover/.config/pubtools-pulp/fake.yaml
    units:
    - _class: FileUnit
      content_type_id: iso
      path: repos.json
      repository_memberships:
      - redhat-maintenance
      sha256sum: ebd05e2128ae0a929bc7163c0457830d3d89d0b758c9aa3f93d087238e125c96
      size: 126

I can also manually modify the state file to simulate any desired
conditions.