tests: fix a non-forwards-compatible test
In tests outside of pubtools-pulplib, it is a bit risky to compare
expected results against precise *Unit instances. The problem is that if
pubtools-pulplib gains support for any new fields, and the fake client
sets those fields to non-default values, then the test would start to
fail as the objects would no longer be precisely equal.

The effect is that such a test may not be forwards-compatible with
newer versions of pubtools-pulplib.

Fix it up by not comparing details we don't care about. In this specific
case, we're only trying to do a sanity check of modulemds we've
uploaded, so we don't need to check every field.