refactor: make @step decorator support generators
I'm currently working on pubtools-pulp-push and finding that I would
like to write some push steps as generators. For example, the
"Calculate checksums" @step can accept items without checksums as input,
and yield items with checksums as output, without having to eagerly load
all items.

@step decorator did not formerly support this; add support for it here.

This is not currently used from any task. I've split this refactor off
from the main pubtools-pulp-push work to reduce the size of the upcoming
PR for that.