push: support the remainder of File mutable fields
version, display_order fields can now be passed onto Pulp, fixing
one of the TODOs. With the latest version of pubtools-pulplib, this will
result in the setting of ud_file_release_mappings_2 repo note during
publish.

(note the comment's reference to content ISOs is outdated, as we decided
that content ISOs won't be ported over to this code.)