Refactor push workflow into loosely coupled phases [RHELDST-8331]
Each step in the push workflow was previously implemented via
generators. That approach limits concurrency as much of the push
logic still takes place in the main thread which can do only one
thing at a time. For example, it is ideally possible for the "calculate
checksums" step to continue calculating checksums of files while the
"query pulp" step is already issuing Pulp queries, but that couldn't be
done with the generator approach.

This commit fixes the problem by refactoring each step of the workflow
into a "phase", which runs on its own thread and communicates with
previous/later phases via queues. It brings several advantages:

- removes unnecessary sources of blocking, as each phase runs at its own
  pace without directly blocking the others

- smarter batching of items becomes possible

- new code is arguably more readable since each phase has clearly
  delineated inputs and outputs

- as queue sizes can be tracked, the progress across phases can be
  monitored and visualized - useful for long-running pushes. (Not
  included in this commit but coming soon)

One of the downsides is that a specific start/end order of each executed
step can no longer be guaranteed. For these reason, some log-based tests
had to be dropped.