Add progress tracking onto push phases [RHELDST-8331]
Track queue events to get an idea of the progress of each phase.
For each phase, by looking at how many items have been put onto the
queue and how many have been taken out, we can figure out the overall
progress of the phase.

This progress info will be logged at a configurable interval as push runs
(every 5 minutes by default). The logging includes both a visual
component (progress bars) and structured metrics which can be gathered
into JSONL logs.

The primary goal of this logging is to make it clear which of the phases
acts as a bottleneck and could benefit from performance improvements or
tuning.