push: parallelize checksum calculation
Testing shows this step to be a bottleneck. Fix the TODO about
parallelizing the checksum calculation. Also make it possible for the
phase to pass through objects which already have checksums as quickly as
possible, so that later phases can start processing those items
earlier.

Some features were added to Phase base class to support this. Other
phases are likely to get similar refactors.