Improve pipelining of upload, update, query & associate push phases
As a holdover from the older generator-based implementation, the logic
in these phases was written in a way such that no items would be output
until all items had been received on the input. This unnecessarily
blocks later phases.

Rewrite them to remove the unnecessary blocking. In each case a refactor
is done so that the phase is able to make use of put_future_output(s),
similarly as done for the checksum phase in fa8cedbd84.