push: fix an error handling mistake which could swallow exceptions
Errors can occur asynchronously when put_future_output is used.
A comment here reasoned that this should not be logged in the error
handling callback because re-raising the exception will cause it
to propagate elsewhere and appropriate logging will happen there.

That reasoning was not correct for all cases. When the error was
re-raised, we'd be putting the phase into a state where __get_input
would raise PhaseInterrupted and __future_puts would raise the original
exception. There is no guarantee which one happens first, as it depends
how the phase implements run(). If the former happened first, the
original exception which caused the interruption would be silently lost.

Update the error handling so that the exception will always be logged in
this scenario.