push: add checks on the outcome of item uploads & updates
When we upload or update an item in Pulp, even if the Pulp API says it
succeeded, we should still double-check the new state of the item to
verify that it had the desired impact. (i.e. if we uploaded an item then
it should be in at least one repo; if we updated an item then it should
not require any further updates.

Fixes a couple of TODOs.