push: reduce modulemd-related blocking in associate
We want to ensure that RPMs are always associated into target repos only
after any corresponding modulemds. Previously, the simplistic solution
to this was simply to not associate *anything* until *all* uploads had
completed, but that causes unnecessary blocking in some cases and
potentially wastes a lot of time for pushes which don't even use RPMs or
modules.

With a minor refactor we can do a little better: make associate process
RPMs last, only after all other items have made it into the phase. Fixes
a TODO.

This could still be improved further by tracking information on which
modules/RPMs correspond to each other and in which repos, but the
performance vs complexity trade-off isn't clear right now.