push: randomly choose upload repo for improved performance
Uploading to a repo locks that repo, so always uploading to the first
available repo can end up introducing a lot of unnecessary
serialization. It's best if we ensure the upload repo is uniformly
distributed across the available choices, which can be done (roughly) by
picking one at random. Fixes a TODO.

Note that fix-cves already does the same thing.