Fix a scaling issue with put_future_output
future_puts here was previously constructed using f_and.

Problem: we can potentially end up with thousands of wrapped
futures in that case, and depending on the order they're resolved, this
can cause a stack overflow when a completion callback is invoked.
This would be most common in the case where a slow operation was
followed by thousands of fast operations.

Fix it by rewriting the code to avoid any deep chain of futures like
that. Since f_and was used to get the "cancel futures when no longer
needed" behavior, that behavior now needs to be implemented explicitly.