push: fix attempting to flush UD cache for all-rpm-content
Due to the errata-specific logic of "if unit changes in any repo,
republish all of them", we could end up trying to publish & cache flush
the all-rpm-content repo here if an erratum unit existed in that repo.
Publish wouldn't be a problem as the pulp client is smart enough to
realize that's a no-op, but cache flush would fail.

Add a condition to avoid doing any publish handling for those repos.