push: make collect phase filter out duplicates
If we're running fast enough, we might end up calling update_push_items
with the "same" item at multiple states, e.g. PENDING when we first
discover it and then EXISTS if we were already able to figure out the
item exists in Pulp.

In that case, we should only include the latest version of the item.
It's important because Pub's collector backend will crash otherwise as
it does not expect to see duplicate items like that.