push: fix files potentially being uploaded more than once
When we're used with staged source it's possible for the same file or
RPM to be present in multiple destination dirs. If so, we get one push
item per instance of that file.

The code here did not previously have any special handling for this, so
if the same file was in N dirs then we'd potentially upload it to pulp
N times. Fix it by keeping track of uploaded content and reusing
existing upload futures when possible.