Fix handling of items with FTP destinations
When loading items from Errata Tool, the current design is for push
items to include both Pulp repo IDs and FTP paths in 'dest', e.g.
a realistic value of dest could be:

    ["rhel-7-server-rpms__7Server__x86_64",
     "/ftp/pub/redhat/linux/enterprise/7Server/en/os/SRPMS/"]

In this code we need to deal only with the Pulp repos. In Pub there was
a simple check of the form "keep only the dest items without '/'" but it
was overlooked here. Add the same logic.