Update product_versions on upload of productid
The legacy push code implemented the following side-effect:
when a productid cert is uploaded into a repo, version strings are
extracted from the cert and appended to the product_versions field in
that repo plus other repos with certain matching notes.

That side-effect of productid upload remains important, but was missing
from pubtools-pulp-push. Add it now.