Fix crash on non-RPM push items with no dest
If a push item has no destinations for push, we would get a crash when
trying to figure out the repo for upload.

Fix it by filtering out such items. The reason they were not filtered
previously is because a push with no destination makes sense at least
for pre-push, so that specific case still allows the item to be
processed.

An item with no dest is treated as a warning rather than error to ensure
forward compatibility with new types of items.