Pkgs in advisory are deleted from all repos when no repos provided
Previously, deleting packages in an advisory expected the repos
to be provided in the request. This patch ensures repos are not
mandatory in the request while deleting an advisory. The packages
in the advisory are now deleted from all the repos the advisory is
a part of if no repos are provided in the request.