tuning: slash default queue size significantly [RHELDST-11374]
There are still memory usage issues with large push tasks.
After doing some analysis and testing I found that the size of each
future is much larger than I'd expected, around 3.3KiB each on python2
(and roughly half of that on py3).

The number of futures which will exist at any one time is roughly
expected to be some multiple of this tunable. The 10,000 number had no
specific calculations behind it and was far too optimistic.
The new number of 200 is set according to some rough estimates that a
number of 2000 would end up consuming roughly 200MB of memory for Future
objects, and then cutting it further by a factor of 10 to get a generous
safety margin.

One unit test needs a larger queue size because the test is written in a
way which will deadlock otherwise.