Add missing support for `--skip publish' to push [RHELDST-11547]
This is an important option for some workflows which want to do several
pushes in sequence and only publish for the last one, but support for it
in this command was missed up to now. Add it.

The 'end pre-push' phase is slightly refactored into a general 'end
push' phase, since skipping publish is another reason for push to end
early.