Erratums are always uploaded to all-rpm-content repo
Pulp maintains the pkglist for (erratum_id, repo) pair. If an
erratum is uploaded for an existing erratum, it's overwritten if
repo is the same, else is appended.
Currently, the erratum is uploaded to one of the repos picked
randomly from the destination list. This messed up the pkglist
as the packages that were modified in the new erratum upload did't
reflect in the package list as the list is appended rather than
modified/overwritten.
Hence, this patch ensures that the erratums are always uploaded to
the all-rpm-content repo to have the same (erratum_id, repo) pair
generating a consistent pkglist.