Increase default phase timeout
A timeout must be set for all phases here for the reasons explained in
the comment, even though in practice we don't really expect that timeout
to ever be reached.

Problem: the earlier value worked out to about ~55 hours and it is
genuinely possible for some push tasks to take that long during
execution, resulting in a spurious timeout as happened in Pub task
545951.

Bump it up by 10x. The new value is about 23 days which surely ought to
be enough...