Pin bandit version
We pin the version of Bandit SAST tool to make the CI tests more
predictable. If we used the latest version available without pinning it,
the CI tests could start failing on new Bandit version without any
changes in our codebase.

We pin the Bandit version by adding it to test-requirements.in.