Reduce log spam for cancelled copy tasks
The code here is responsible for logging a summary of a successfully
completed copy. The intent is to do this only when a copy succeeded,
hence the f.exception() check. However, f.exception() itself raises an
exception if the future was cancelled which would generate a lot of
log spam.

Fix it to do the logging only if the future succeeded, per the original
intent.