Revert "Retry all failed associations with up-to-date data"
According to testing in RHELDST-18718, this commit introduced a deadlock
which is proving difficult to track down and fix. A deadlock is a
serious issue which shouldn't be deployed into production, so we should
not leave this sitting in master because it effectively blocks the
release of any other changes in this repo.

This reverts commit 3dcb3a45c78569c447da87899c706ebd8f05bef1.