Remove rpms in the modules from all the provided repos [RHELDST-13651]
Currently, rpms in a module are deleted only from the repos
where the module exists when a module is deleted. This leaves
the rpms from the module in other repos even after module was
deleted e.g. modules are present in main rpm repo but the related
rpms could be in other repos like source, debug etc. On deletion
of the module, rpms from only the main repo were deleted and not
from source, debug etc.
This ensures the rpms are deleted from all the provided repos once
it's established that module is present in one of the provided
repos.