Remove EngID flush [RHELDST-29701]
Previously, during each Flush UD cache step, the Eng ID for each
flushed repository was also flushed. This resulted in repeated
flushes of the same Eng ID.
It turns out that flushing the Eng ID is not necessary at all,
and that just flushing the repos should be enough to keep the
UD up to date.