Flush engproduct ID once [RHELDST-30255]
Previously, the UD cache for an eng product ID was flushed whenever it
was found in a repo. This functionality was removed, as it put heavy
load on the UD cache's API. We later found the product ID should in fact
be flushed.

This change adds back in the UD cache flushing for eng product IDs, but
makes sure each ID is only flushed once, rather than flushed for each
repo it's in.